+++ src/main/java/nl/novi/finalAssignmentBackend/model/ProductModel.java
@@ -5,7 +5,7 @@ package nl.novi.finalAssignmentBackend.model;
 public abstract class ProductModel {
 
 
-    private Integer sellingPrice;
+    private Double sellingPrice;
 
     private Integer originalStock;
 
@@ -25,7 +25,7 @@ public abstract class ProductModel {
         this.yearOfRelease = yearOfRelease;
     }
 
-    private Integer purchasePrice;
+    private Double purchasePrice;
 
 
 
@@ -37,19 +37,19 @@ public abstract class ProductModel {
 
 
 
-    public Integer getSellingPrice() {
+    public Double getSellingPrice() {
         return sellingPrice;
     }
 
-    public void setSellingPrice(Integer price) {
+    public void setSellingPrice(Double price) {
         this.sellingPrice = price;
     }
 
-    public Integer getPurchasePrice() {
+    public Double getPurchasePrice() {
         return purchasePrice;
     }
 
-    public void setPurchasePrice(Integer purchasePrice) {
+    public void setPurchasePrice(Double purchasePrice) {
         this.purchasePrice = purchasePrice;
     }
